requirements.txt
schedule
python-dotenv
requests
websocket-client
flask